(WIP) Reduce arg/param mappings
Use an inverted copy of resource_map to parse rules from iptables-save. This
saves maintaining two copies of the argument to param mappings. Arguments
for required modules and alternative arguments are also split out into
separate hashes.

I was hoping to use resource_map to create rules in general_args as well,
since this new parser makes us agnostic of the ordering in iptables-save.
However all of the arrays in the test fixtures still depend on order.